+++ src/test/java/org/terracotta/offheapstore/CacheTestRoutines.java
@@ -1,4 +1,4 @@
-/* 
+/*
  * Copyright 2015 Terracotta, Inc., a Software AG company.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -36,34 +36,34 @@ public final class CacheTestRoutines {
 
   public static void testCacheEviction(Map<Integer, Integer> cache, CapacityLimitedIntegerStorageEngineFactory engine) {
     Random rndm = new Random();
-    
+
     int toAdd = 100 + rndm.nextInt(90);
-    
+
     engine.setCapacity(toAdd);
-    
+
     for (int i = 0; i < toAdd; i++) {
       cache.put(i, i);
 
       Assert.assertEquals(i + 1, cache.size());
       for (int j = 0; j <= i; j++) {
-        Assert.assertTrue(cache.get(j).intValue() == j);
+        Assert.assertTrue(cache.get(j) == j);
       }
     }
-    
+
     int toEvict = 100 + rndm.nextInt(90);
-    
+
     for (int i = toAdd; i < toEvict + toAdd; i++) {
       cache.put(i, i);
 
       Assert.assertEquals(toAdd, cache.size());
-      Assert.assertTrue(cache.get(i).intValue() == i);
+      Assert.assertTrue(cache.get(i) == i);
     }
   }
 
   public static void testCacheEvictionMinimal(Map<Integer, Integer> cache) {
     for (int i = 0; i < 200; i++) {
       cache.put(i, i);
-      Assert.assertTrue(cache.get(i).intValue() == i);
+      Assert.assertTrue(cache.get(i) == i);
     }
   }
 
@@ -91,22 +91,22 @@ public final class CacheTestRoutines {
         Assert.assertEquals(keySetCopy, cache.keySet());
       }
     }
-    
+
     assertThat(cache.isEmpty(), is(false));
 
     Set<Integer> keySetCopy = new HashSet<Integer>(cache.keySet());
-    
+
     for (Integer k : keySetCopy) {
       if (AbstractOffHeapMapIT.doFill(cache, k, new byte[(k % 1024) + 1]) == null) {
         if (cache.containsKey(k)) {
-          Assert.assertEquals((k % 1024) + 1, cache.get(k).length); 
+          Assert.assertEquals((k % 1024) + 1, cache.get(k).length);
         }
       } else {
         Assert.assertTrue(cache.containsKey(k));
-        Assert.assertEquals((k % 1024) + 1, cache.get(k).length); 
+        Assert.assertEquals((k % 1024) + 1, cache.get(k).length);
       }
     }
-  }  
+  }
 
   public static void testComputeEvictionBehavior(Map<Integer, byte[]> cache) {
     for (int i = 0; i < 100000; i++) {
